* Source/externs.m: Change predefined colour values to CGFloat.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36037 72102866-910b-0410-8b05-ffd578937521

+++ Source/externs.m
@@ -141,11 +141,11 @@ NSString *NSPatternColorSpace = @"NSPatternColorSpace";
 NSString *NSCustomColorSpace = @"NSCustomColorSpace";
 
 // NSColor Global gray values
-const float NSBlack = 0;
-const float NSDarkGray = .333;
-const float NSGray = 0.5;
-const float NSLightGray = .667;
-const float NSWhite = 1;
+const CGFloat NSBlack = 0;
+const CGFloat NSDarkGray = .333;
+const CGFloat NSGray = 0.5;
+const CGFloat NSLightGray = .667;
+const CGFloat NSWhite = 1;
 
 // NSColor notification
 NSString *NSSystemColorsDidChangeNotification =
